cloud_funcs/requirements.txt
functions-framework==3.*
Flask==3.0.3
google-cloud-bigquery==3.11.4
rapidfuzz==3.13.0